Rescuecore2: Moved version0 property types into rescuecore2.worldmodel.properties and tinkered with EntityType and PropertyType management.
BooleanProperty, IntProperty and so on are all reusable across different implementations of entities so they have been moved into the main package.

EntityType is now an interface; added PropertyType. AbstractEntity is parameterized on the subclass of EntityType that it returns so users don't need to downcast so much.

Changed the EntityConstants class into an enum called RescueEntityType.